all: Decouple pkg/errors tests
